final ediitt

+++ Admin/assets/js/addProducts.js
@@ -50,6 +50,13 @@ function addProduct(imgBase64) {
   products.push(product);
   localStorage.setItem("productsList", JSON.stringify(products));
   // count++;
+  Swal.fire({
+    position: 'top-center',
+    icon: 'success',
+    title: 'Your work has been saved',
+    showConfirmButton: false,
+    timer: 2000
+  })  
 }
 
 function displayData() {
@@ -104,8 +111,25 @@ function clearForm() {
 }
 
 function deleteProduct(index){
-  products.splice(index, 1);
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+  }).then((DisplayProductForAdmin) => {
+if (DisplayProductForAdmin.isConfirmed) { 
+  products.splice(index, 1);  
   localStorage.setItem("productsList",JSON.stringify(products));
   displayData();
+  Swal.fire(
+    'Deleted!',
+    'Your product has been deleted.',
+    'success'
+          )
+        }
+      })
 }
